# Optimize `StringName` creation by using `LocalVector` instead of linked lists, reducing cache misses.