@YYF233333 2025-03-19
> Unfortunately, `StringName::_Data` cannot be fully localized because the pointer of existing `StringName` instances must not be invalidated.

Speaking of this, I think `PagedArray`(`core/templates/paged_array.h`) may be relevant. Happened to come across it today and based on its description it won't move its elements.