@Ivorforce 2025-03-19
> Speaking of this, I think PagedArray(core/templates/paged_array.h) may be relevant. Happened to come across it today and based on its description it won't move its elements.

Yeah, I can imagine `PagedArray` could help performance further. Perhaps we could take a look at that some time too!