@Ivorforce 2025-03-19
> Maybe it would be worth tinkering with the buckets size, the whole 16bits, aka 65k buckets seems a lot, but hard to tell without data.

Agreed. Big buckets were a bottleneck before this PR, but since that's now alleviated we can definitely rethink the bucket sizes.