@aaronp64 2025-03-19
Could you try including removing Nodes from the tree in your c++ benchmark?  I tried benchmarking in gdscript with the below - smaller difference is expected with the gdscript overhead, but what seems odd is the removal is consistently slower for me.  I'm wondering how much of the difference comes from using `LocalVector::push_back()` to add new data to the end, while before new data was added to the beginning of the list.

I think there's two main ways StringNames are involved when creating new Nodes.  When new nodes are added to the tree, they get a unique name assigned if they don't already have one.  When adding a ton of Nodes to the tree at the same time, it basically floods the StringName table with these unique names, resulting in the lists/LocalVectors becoming longer.  There's also the StringNames from the GDVIRTUAL macros on Node - "_process", "_physics_process", etc.  These are created (looked up, not reallocated) for each created Node, even though they all have the same value (maybe these could be static in the future?).  Before, since the node names were added to the beginning of the lists, the macro strings kept getting pushed back, making them take longer to look up on node creation.  Now that the macro strings stay at the beginning, they're found faster, and Node creation is faster.  But that would also mean the node names are further back in the LocalVector, so freeing nodes and deleting the name string has to look further and takes longer.

In gdscript at least, the add speeds up by more than the remove slows down for large numbers of nodes, so the tradeoff seems to be worth it, but would be good to get a better idea what's happening on the c++ side.

```gdscript
func _ready() -> void:
	test_add_remove(50000)
	test_add_remove(500000)
	test_add_remove(1000000)

func test_add_remove(p_count : int):
	var start := Time.get_ticks_msec()
	for i in p_count:
		add_child(Node.new())
	var end := Time.get_ticks_msec()
	print("%s %d: %dms" % ["add", p_count, end - start])
	start = Time.get_ticks_msec()
	for child in get_children():
		child.free()
	end = Time.get_ticks_msec()
	print("%s %d: %dms" % ["rem", p_count, end - start])
```

old:
```
add 50000: 132ms
rem 50000: 68ms
add 500000: 1469ms
rem 500000: 730ms
add 1000000: 3760ms
rem 1000000: 1778ms
```

new
```
add 50000: 139ms
rem 50000: 69ms
add 500000: 1371ms
rem 500000: 796ms
add 1000000: 3037ms
rem 1000000: 1945ms
```