@Ivorforce 2025-03-19
> Could you try including removing Nodes from the tree in your c++ benchmark? I tried benchmarking in gdscript with the below - smaller difference is expected with the gdscript overhead, but what seems odd is the removal is consistently slower for me. 

Thank you for testing!
I definitely should have tested `Node` removal from the start. I can confirm it's somewhat slower, but not as slow on my end as on yours:

```c++
	Node *parent = memnew(Node);

	{
		auto t0 = std::chrono::high_resolution_clock::now();
		for (int i = 0; i < 2000000; i ++) {
			Node *child = memnew(Node);
			parent->add_child(child);
		}
		auto t1 = std::chrono::high_resolution_clock::now();
		std::cout << std::chrono::duration_cast<std::chrono::milliseconds>(t1 - t0).count() << "ms\n";
	}
	{
		Array children = parent->get_children();
		auto t0 = std::chrono::high_resolution_clock::now();
		for (int i = 0; i < 2000000; i ++) {
			parent->remove_child((Node *)(Object *)children.get(i));
		}
		auto t1 = std::chrono::high_resolution_clock::now();
		std::cout << std::chrono::duration_cast<std::chrono::milliseconds>(t1 - t0).count() << "ms\n";
	}
```

Yielding
- `3630ms`, `675ms` on master
- `2677ms`, `778ms` on this PR

(adding `27%` /  `1000ms` faster, removing `15%` / `103ms` slower)

I'm not sure why your test ended up so much worse. Are you comparing against 4.4, or against 3d9b05ad4aab12c86fd83e4cb5b212e550cb8372?