@aaronp64 2025-03-20
> Some difference is to be expected, because removing from the list when you have the element pointer requires only 2 or 3 dereferences, while removing from LocalVector requires that but also searching for the entry first.


That makes sense!  I was thinking of both versions looping through in `unref` after looking at the new code, but the old code didn't have that loop.  Might be something to watch out for if we end up reducing the number of buckets.  Also, maybe it would be possible to store the LocalVector index in _data (if we can keep it updated after `remove_at_unordered`), so it could jump straight there?  That sounds like it might add an extra pointer to follow for each delete though, not sure if it'd actually help.

I was testing with ef1153baf and the changes applied from there.  It varies a bit from run to run, but I think the percentage differences are similar for the larger set, maybe a bit smaller for going through gdscript.  Overall, even with removes being slower, the combined add+remove definitely seems to scale up better with this change.