@Ivorforce 2025-03-20
> LocalVector index in _data (if we can keep it updated after remove_at_unordered)

I was about to say that would cost as much as it brings, but with remove_at_unordered we would actually only have to update a single Data's value! I'll test that.