@Ivorforce 2025-04-02
@aaronp64 Your suggestion works great! Caching the entry's index so removing entries does not need to search the array seems to eliminate any measurable 'removal' differences for me.

I tested your code again to make sure it addresses the problem. The differences don't seem to be as much as for you, but it is measurable:
```
add 50000: 63ms
rem 50000: 36ms
add 500000: 678ms
rem 500000: 379ms
add 1000000: 1791ms
rem 1000000: 1025ms
```
(master)
```
add 50000: 69ms
rem 50000: 32ms
add 500000: 667ms
rem 500000: 393ms
add 1000000: 1518ms
rem 1000000: 1069ms
```
(this PR)