@aaronp64 2025-04-09
I reran the gdscript comparison with more recent master.  I was curious to see how changing the GDVIRTUAL StringNames to static in https://github.com/godotengine/godot/pull/104264 would change the performance impact on `Node` creation from this PR - still seeing similar improvements:

master
```
add 50000: 135ms
rem 50000: 76ms
add 500000: 1329ms
rem 500000: 826ms
add 1000000: 3226ms
rem 1000000: 1953ms
add 2000000: 8642ms
rem 2000000: 4057ms
```

PR
```
add 50000: 138ms
rem 50000: 72ms
add 500000: 1269ms
rem 500000: 833ms
add 1000000: 2784ms
rem 1000000: 2086ms
add 2000000: 6362ms
rem 2000000: 4186ms
```