@Ivorforce 2025-04-28
My other recent `StringName` optimizations have somewhat changed the playing field. In @aaronp64's test from above, there is little improvement over `master`:

```
add 50000: 52ms
rem 50000: 39ms
add 500000: 522ms
rem 500000: 444ms
add 1000000: 1306ms
rem 1000000: 1111ms
```
(master)

```
add 50000: 57ms
rem 50000: 39ms
add 500000: 548ms
rem 500000: 430ms
add 1000000: 1292ms
rem 1000000: 1122ms
```
(this PR)

In my own c++ test, I can still measure a ~8% improvement, but it's far less interesting than before.
I'll have to re-profile to make sure it's still worth it. It may make sense to further tune the current configuration as well before switching out the implementation entirely.