@Ivorforce 2025-05-13
I've once again run some tests, this time with a bit more sophistication with different amounts of `StringName`:
<details>
<summary>Test Code</summary>

```c++
	size_t N = 20000000;
	for (size_t array_size : { 1, 8, 64, 1000, 10000, 100000, 1000000 }) {
		{
			size_t total_ns = 0;
			for (size_t j = 0; j < N / array_size; ++j) {
				LocalVector<StringName> names;
				names.resize(array_size);

				auto t0 = std::chrono::high_resolution_clock::now();
				for (size_t i = 0; i < array_size; i ++) {
					names[i] = StringName(String::num_int64(i));
				}
				auto t1 = std::chrono::high_resolution_clock::now();
				total_ns += std::chrono::duration_cast<std::chrono::nanoseconds>(t1 - t0).count();
			}

			std::cout << "construct " << array_size << std::endl;
			std::cout << total_ns / 1000 / 1000 << "ms\n";
		}
		{
			size_t total_ns = 0;
			for (size_t j = 0; j < N / array_size; ++j) {
				LocalVector<StringName> names;
				names.resize(array_size);

				for (size_t i = 0; i < array_size; i ++) {
					names[i] = StringName(String::num_int64(i));
				}
				auto t0 = std::chrono::high_resolution_clock::now();
				for (size_t i = 0; i < array_size; i ++) {
					names[i] = StringName();
				}
				auto t1 = std::chrono::high_resolution_clock::now();
				total_ns += std::chrono::duration_cast<std::chrono::nanoseconds>(t1 - t0).count();
			}

			std::cout << "destruct " << array_size << std::endl;
			std::cout << total_ns / 1000 / 1000 << "ms\n";
		}
	}
```
</details>

I tested current master, this branch (rebased), and also this branch with just `2^14` buckets (1/4th).
This gave the following results:
<details>
<summary>Detailed Results</summary>

```
PR 2^16
construct 1
903ms
destruct 1
738ms
construct 8
640ms
destruct 8
390ms
construct 64
666ms
destruct 64
385ms
construct 1000
861ms
destruct 1000
388ms
construct 10000
844ms
destruct 10000
403ms
construct 100000
1000ms
destruct 100000
500ms
construct 1000000
1472ms
destruct 1000000
970ms


PR 2^14
construct 1
902ms
destruct 1
736ms
construct 8
641ms
destruct 8
393ms
construct 64
722ms
destruct 64
386ms
construct 1000
1658ms
destruct 1000
409ms
construct 10000
1006ms
destruct 10000
426ms
construct 100000
1155ms
destruct 100000
536ms
construct 1000000
1807ms
destruct 1000000
869ms

master
construct 1
909ms
destruct 1
716ms
construct 8
632ms
destruct 8
377ms
construct 64
661ms
destruct 64
436ms
construct 1000
772ms
destruct 1000
433ms
construct 10000
842ms
destruct 10000
484ms
construct 100000
1027ms
destruct 100000
548ms
construct 1000000
2103ms
destruct 1000000
648ms
```

In short, I've found that this PR is still better than `master` across the board, though not as much as before my other optimizations:
- `1 StringName:` ~same
- `8 StringName:` 2% worse
- `64 StringName:` 4% better
- `1000 StringName:` 4% better
- `10000 StringName:` 6% better
- `100000 StringName:` 5% better
- `1000000 StringName:` 12% better

The differences are way worse if i decrease the number of buckets (this PR with 2^14 buckets, compared to master):
- `1 StringName:` ~same
- `8 StringName:` 2% worse
- `64 StringName:` 1% worse
- `1000 StringName:` 70% worse
- `10000 StringName:` 8% worse
- `100000 StringName:` 7% worse
- `1000000 StringName:` 3% better

At this point, I'm not sure if this PR is still worth it. `10%` improvement isn't _nothing_, but it's also not all that exciting. Considering it's somewhat of a risk to change the implementation in such a drastic way, it may be more interesting to explore another approach entirely.
(this conclusion comes with the caveat that these are benchmarks, and this may affect actual programs differently due to cache contention)

For anyone that's been following this PR, do not fret - my previous PRs have already sped up `StringName` significantly, so we should still be seeing most of the benefits proclaimed here initially :)